requirements.txt
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.3.2
gitpython==3.1.32
pygithub==1.59.0
python-gitlab==3.15.0
atlassian-python-api==3.41.1
waitress==2.1.2